Allow to run Calls slash commands
